refactor(api, libs): Centralize and standardize DTOs for reusability
This commit centralizes `CreateVariableDto`, `UpdateVariableDto`, and introduces
new DTOs for creating and updating execution nodes and edges within the
`caidense-reasoning` library. This streamlines DTO management and promotes
reusability across different modules.

Previously, `CreateVariableDto` and `UpdateVariableDto` were duplicated
or scattered in various API modules. By moving them to the `common/dto`
directory in `libs/caidense-reasoning`, they can now be consistently
imported and used wherever variable data needs to be handled.

Additionally, dedicated DTOs for `CreateExecutionEdgeDto`,
`UpdateExecutionEdgeDto`, and `UpdateExecutionNodeDto` have been created
under `libs/caidense-reasoning/src/edge/dto` and `libs/caidense-reasoning/src/node/dto`
respectively. This provides clear, versioned interfaces for creating and
updating these critical components of reasoning flows.

Impacted areas include:
- **API Modules**: `api/src/modules/prompt/text` and `api/src/modules/reasoning/thinking`
  now import DTOs from the centralized `caidense-reasoning` library.
- **DTO Definitions**: Old, duplicated DTOs in `api/src/modules/base/dto`
  have been removed.
- **Consistency**: Ensures a single source of truth for DTO definitions,
  reducing potential for discrepancies and simplifying future maintenance.
