+++ api/src/modules/reasoning/thinking/dto/create-thinking.dto.ts
@@ -1,6 +1,7 @@
-import { CreateVariableDto } from '@/modules/base/dto/create-variable.dto';
-import { CreateReasoningNodeDto } from '@/modules/reasoning/node/dto/create-node.dto';
 import { DocumentStatus } from '@caidense/reasoning/common/common.interface';
+import { CreateVariableDto } from '@caidense/reasoning/common/dto/common.dto';
+import { CreateExecutionEdgeDto } from '@caidense/reasoning/edge/dto/create-edge.dto';
+import { CreateExecutionNodeDto } from '@caidense/reasoning/node/dto/create-node.dto';
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import {
@@ -12,59 +13,6 @@ import {
   ValidateNested
 } from 'class-validator';
 
-
-export class CreateReasoningThinkingEdgeDto {
-
-  @ApiProperty({
-    description: 'The ID of the source node',
-    example: 'source-node-id',
-  })
-  @IsString()
-  source: string;
-
-  @ApiProperty({
-    description: 'The specific handle/port ID on the source node where the edge starts.',
-    example: 'true',
-  })
-  @IsOptional()
-  @IsString()
-  sourceHandle?: string;
-
-  @ApiProperty({
-    description: 'The ID of the target node',
-    example: 'target-node-id',
-  })
-  @IsString()
-  target: string;
-
-  @ApiProperty({
-    description: 'The specific handle/port ID on the target node where the edge ends.',
-    example: 'false',
-  })
-  @IsOptional()
-  @IsString()
-  targetHandle?: string;
-
-  @ApiProperty({
-    description: 'The type of the edge, which is optional.',
-    example: 'default',
-  })
-  @IsOptional()
-  @IsString()
-  type?: string;
-
-  @ApiProperty({
-    description: 'The label of the edge, which is optional.',
-    example: 'Edge Label',
-  })
-  @IsOptional()
-  @IsString()
-  label?: string;
-
-  // Note: Fields like 'status', 'createdAt', 'updatedAt', or internal IDs (_id)
-  // are typically managed by the backend service and the database,
-  // and thus are not included in the DTO used for *creating* the resource.
-}
 
 export class CreateReasoningThinkingDto {
 
@@ -85,19 +33,19 @@ export class CreateReasoningThinkingDto {
 
   @ApiProperty({
     description: 'An array of nodes within the thinking flow diagram.',
-    type: [CreateReasoningNodeDto],
+    type: [CreateExecutionNodeDto],
     isArray: true,
   })
-  @Type(() => CreateReasoningNodeDto)
-  nodes: CreateReasoningNodeDto[];
+  @Type(() => CreateExecutionNodeDto)
+  nodes: CreateExecutionNodeDto[];
 
   @ApiProperty({
     description: 'An array of edges connecting nodes within the thinking flow diagram.',
-    type: [CreateReasoningThinkingEdgeDto],
+    type: [CreateExecutionEdgeDto],
     isArray: true,
   })
-  @Type(() => CreateReasoningThinkingEdgeDto)
-  edges: CreateReasoningThinkingEdgeDto[];
+  @Type(() => CreateExecutionEdgeDto)
+  edges: CreateExecutionEdgeDto[];
 
   @ApiProperty({
     description: 'An array of input variables defined for the thinking flow.',
