+++ libs/caidense-reasoning/src/node/dto/update-node.dto.ts
@@ -1,27 +1,25 @@
-import { UpdateVariableDto } from '@caidense/reasoning/common/dto/common.dto';
 import { ApiProperty } from '@nestjs/swagger';
+import { IsString, IsOptional, IsObject, IsArray, ValidateNested } from 'class-validator';
 import { Type } from 'class-transformer';
-import { IsArray, IsObject, IsOptional, IsString, ValidateNested } from 'class-validator';
+import { UpdateVariableDto } from '@caidense/reasoning/common/dto/common.dto';
 
 
 /**
- * Defines the data structure for updating an existing Reasoning Node document.
- * All fields are optional for partial updates.
+ * Defines the data structure for creating a new Reasoning Node document.
+ * Excludes backend-generated fields like _id and timestamps.
  */
-export class UpdateReasoningNodeDto {
+export class UpdateExecutionNodeDto {
   @ApiProperty({
     description: 'The type of the node, determining its function and behavior in the flow.',
     example: 'systemLogic',
-    required: false
   })
-  @IsOptional()
   @IsString()
-  type?: string;
+  type: string;
 
   @ApiProperty({
     description: 'An optional label or name displayed on the node.',
     required: false,
-    example: 'Updated Label',
+    example: 'Calculate & Compare Sum',
   })
   @IsOptional()
   @IsString()
@@ -30,7 +28,7 @@ export class UpdateReasoningNodeDto {
   @ApiProperty({
     description: 'Optional configuration object specific to the node type.',
     required: false,
-    example: { threshold: 150 },
+    example: { logicType: 'sumAndCompare', threshold: 100 },
   })
   @IsOptional()
   @IsObject()
@@ -51,7 +49,7 @@ export class UpdateReasoningNodeDto {
   @ApiProperty({
     description: 'Optional script code to be executed by this node type.',
     required: false,
-    example: 'let sum = a + b; return sum > 150;'
+    example: 'let sum = a + b; return sum > 100;'
   })
   @IsOptional()
   @IsString()
